@@ -1,29 +1,29 @@
 name: time_spent
-description: A new Flutter project.
+description: Track the time you spend on projects.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
 
 dependencies:
-  bloc: ^3.0.0
-  equatable: ^1.0.0
-  flushbar: <1.9.1
   flutter:
     sdk: flutter
-  flutter_bloc: ^3.2.0
-  intl: ^0.16.1
-  moor: ^2.4.2
-  moor_ffi: ^0.4.0
-  path: ^1.6.4
-  path_provider: ^1.6.5
-  provider: ^4.0.4
+  another_flushbar: ^1.10.11
+  equatable:
+  flutter_bloc: ^6.1.2
+  intl:
+  moor:
+  path: ^1.7.0
+  path_provider: ^1.6.27
+  provider: ^4.3.3
+  sqlite3_flutter_libs: ^0.4.0+1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.8.1
-  moor_generator: ^2.4.0
+  build_runner: ^1.11.1
+  moor_generator:
+  pedantic:
 
 
 flutter:
