@@ -2,17 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../data/db/database.dart';
-import '../services/dao/rushes_dao.dart';
 import '../utils/extensions.dart';
 import '../widgets/back_button.dart';
 import 'month_list.dart';
 
 class ListPage extends StatelessWidget {
+  const ListPage({Key key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Material(
       child: Container(
-        decoration: BoxDecoration(
+        decoration: const BoxDecoration(
           gradient: LinearGradient(
             begin: Alignment.topRight,
             end: Alignment.bottomLeft,
@@ -25,26 +26,22 @@ class ListPage extends StatelessWidget {
             const BackWidget(),
             Expanded(
               child: StreamBuilder<Map<String, List<Rush>>>(
-                stream: RushesDao(Provider.of<Database>(context))
-                    .watchMonthlySortedRushes,
+                stream: context.read<DB>().rushesDao.watchMonthlySortedRushes,
                 builder: (context, snapshot) {
                   if (snapshot.hasData) {
                     if (snapshot.data.isEmpty) {
-                      return Center(
-                        child: Text('No data...'),
-                      );
+                      return const Center(child: Text('No data...'));
                     }
                     return ListView.builder(
                       itemCount: snapshot.data.length,
                       itemBuilder: (_, i) => MonthCard(
-                        month: [...snapshot.data.keys][i],
-                        rushes: [...snapshot.data.values][i],
+                        month: snapshot.data.keys.elementAt(i),
+                        rushes: snapshot.data.values.elementAt(i),
                       ),
                     );
-                  } else
-                    return Center(
-                      child: CircularProgressIndicator(),
-                    );
+                  } else {
+                    return const Center(child: CircularProgressIndicator());
+                  }
                 },
               ),
             ),
@@ -56,14 +53,14 @@ class ListPage extends StatelessWidget {
 }
 
 class MonthCard extends StatelessWidget {
-  final String month;
-  final List<Rush> rushes;
-
   const MonthCard({
     Key key,
     @required this.month,
     @required this.rushes,
   }) : super(key: key);
+
+  final String month;
+  final List<Rush> rushes;
 
   @override
   Widget build(BuildContext context) {
@@ -75,6 +72,12 @@ class MonthCard extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.all(8),
       child: InkWell(
+        onTap: () => Navigator.push(
+          context,
+          MaterialPageRoute(
+            builder: (_) => MontListPage(month: month),
+          ),
+        ),
         child: Card(
           color: Colors.indigo[400],
           child: Padding(
@@ -84,14 +87,14 @@ class MonthCard extends StatelessWidget {
               children: <Widget>[
                 Row(
                   children: <Widget>[
-                    Icon(
+                    const Icon(
                       Icons.calendar_today,
                       color: Colors.white,
                     ),
-                    SizedBox(width: 8),
+                    const SizedBox(width: 8),
                     Text(
                       month,
-                      style: TextStyle(
+                      style: const TextStyle(
                         fontStyle: FontStyle.normal,
                         fontSize: 18,
                       ),
@@ -100,19 +103,13 @@ class MonthCard extends StatelessWidget {
                 ),
                 Text(
                   total.pretty(),
-                  style: TextStyle(
+                  style: const TextStyle(
                     fontStyle: FontStyle.normal,
                     fontSize: 18,
                   ),
                 )
               ],
             ),
-          ),
-        ),
-        onTap: () => Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (_) => MontListPage(month: month),
           ),
         ),
       ),
